Сертификаты
изменил стиль ссылок чуть
продолжаем переходить на mysqli

+++ payment_del_f.php
@@ -40,23 +40,39 @@
 
 
                         //Ну вроде все норм, поехали всё обновлять/сохранять
-                        connectDB();
+                        $msql_cnnct = ConnectToDB2 ();
 
                         $payed = $invoice_j[0]['paid'] - $payment_j[0]['summ'];
 
                         //Обновим цифру оплаты в наряде
                         $query = "UPDATE `journal_invoice` SET `paid`='$payed', `status`='0', `closed_time`='0'  WHERE `id`='{$_POST['invoice_id']}'";
-                        mysql_query($query) or die(mysql_error() . ' -> ' . $query);
+                        $res = mysqli_query($msql_cnnct, $query) or die(mysqli_error($msql_cnnct).' -> '.$query);
 
-                        $debited = $client_balance['debited'] - $payment_j[0]['summ'];
+                        if ($payment_j[0]['type'] != 1) {
+                            $debited = $client_balance['debited'] - $payment_j[0]['summ'];
+                        }else{
+                            //Вернем в работу сертификат
+                            $time = date('Y-m-d H:i:s', time());
 
-                        //Обновим потраченное в балансе
-                        $query = "UPDATE `journal_balance` SET `debited`='$debited'  WHERE `client_id`='{$_POST['client_id']}'";
-                        mysql_query($query) or die(mysql_error() . ' -> ' . $query);
+                            if ($payment_j[0]['cert_id'] != 0) {
+                                //Ищем сертификат
+                                $cert_j = SelDataFromDB('journal_cert', $payment_j[0]['cert_id'], 'id');
+                                if ($cert_j != 0) {
+                                    //Обновим потраченное в балансе
+                                    $query = "UPDATE `journal_cert` SET `last_edit_time`='{$time}', `last_edit_person`='{$_SESSION['id']}', `debited`='".($cert_j[0]['debited'] - $payment_j[0]['summ'])."', `closed_time`='0000-00-00 00:00:00', `status`='7'  WHERE `id`='{$payment_j[0]['cert_id']}'";
+                                    $res = mysqli_query($msql_cnnct, $query) or die(mysqli_error($msql_cnnct).' -> '.$query);
+                                }
+                            }
+                        }
+                        if ($payment_j[0]['type'] != 1) {
+                            //Обновим потраченное в балансе
+                            $query = "UPDATE `journal_balance` SET `debited`='$debited'  WHERE `client_id`='{$_POST['client_id']}'";
+                            $res = mysqli_query($msql_cnnct, $query) or die(mysqli_error($msql_cnnct).' -> '.$query);
+                        }
 
                         //Удаляем оплату из БД
                         $query = "DELETE FROM `journal_payment` WHERE `id`='{$_POST['id']}'";
-                        mysql_query($query) or die(mysql_error() . ' -> ' . $query);
+                        $res = mysqli_query($msql_cnnct, $query) or die(mysqli_error($msql_cnnct).' -> '.$query);
 
                         //Обновим общий
                         calculateDebt($_POST['client_id']);
